requirements.txt
groq>=0.9.0
duckduckgo-search>=6.1.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0